+++ programs/direnv/default.nix
@@ -1,0 +1,8 @@
+# https://github.com/direnv/direnv
+{
+  programs.direnv = {
+    enable = true;
+    enableZshIntegration = true;
+    nix-direnv.enable = true;
+  };
+}