+++ programs/fd/default.nix
@@ -1,0 +1,6 @@
+# https://github.com/sharkdp/fd
+{
+  programs.fd = {
+    enable = true;
+  };
+}