+++ programs/fzf/default.nix
@@ -1,3 +1,4 @@
+# https://github.com/junegunn/fzf
 {
   programs.fzf = {
     enable = true;