+++ programs/gh/default.nix
@@ -1,0 +1,9 @@
+# https://github.com/cli/cli
+{
+  programs.gh = {
+    enable = true;
+    settings = {
+      git_protocol = "ssh";
+    };
+  };
+}