+++ programs/jq/default.nix
@@ -1,0 +1,6 @@
+# https://github.com/jqlang/jq
+{
+  programs.jq = {
+    enable = true;
+  };
+}