+++ programs/man/default.nix
@@ -1,0 +1,7 @@
+# https://gitlab.com/man-db/man-db
+{
+  programs.man = {
+    enable = true;
+    generateCaches = true;
+  };
+}